requirements.txt
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
passlib[bcrypt]
PyJWT
